Hardcode PrimaryGenerator engine to mt19937
