+++ app/celer-sim/Runner.cc
@@ -364,7 +364,7 @@ void Runner::build_events(RunnerInput const& inp)
     if (inp.primary_gen_options)
     {
         std::mt19937 rng;
-        auto generate_event = PrimaryGenerator<std::mt19937>::from_options(
+        auto generate_event = PrimaryGenerator::from_options(
             core_params_->particle(), inp.primary_gen_options);
         read_events([&generate_event, &rng] { return generate_event(rng); });
     }
